requirements.txt
requests
beautifulsoup4
playwright
gspread
oauth2client
openai
python-dotenv